+++ apps/admin_app/lib/admin_app_web/templates/layout/sidebar.html.eex
@@ -13,7 +13,7 @@
     </div>
   </li>
   <li>
-    <a class="nav-link" href=<%= "/dashboard?from=#{date_days_before(30)}&to=#{date_today()}" %>>
+    <a class="nav-link" href=<%= "/dashboard?from=#{Helpers.date_days_before(30)}&to=#{Helpers.date_today()}" %>>
     <i class="fa fa-clipboard-list" aria-hidden="true"></i> <span> Dashboard</span>
     </a>
   </li>
